Make Todo::cmp take into consideration the completed state of the todo
Completed todos appear at the bottom of the list in the given
block they are associated with. For example, incomplete A
priority tasks will appear above complete A priority tasks, but
complete A priority tasks will appear before incomplete B
priority tasks.

There is a lot of duplicate logic here, I think it is prime for
refactoring.